quant_engine/requirements.txt
fastapi
uvicorn
hyperliquid
eth-account
python-dotenv
requests